requirements.txt
Pillow
fingerprint_enhancer
pandas
matplotlib
